+++ aliyun-net-sdk-core.Tests/aliyun-net-sdk-core-unit-tests.csproj
@@ -23,6 +23,5 @@
 
   <ItemGroup>
     <ProjectReference Include="..\aliyun-net-sdk-core\aliyun-net-sdk-core.vs2017.csproj" />
-    <ProjectReference Include="..\aliyun-net-sdk-ecs\aliyun-net-sdk-ecs.vs2017.csproj" />
   </ItemGroup>
 </Project>
