+++ aliyun-net-sdk.sln
@@ -7,6 +7,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "aliyun-net-sdk-core.vs2017"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "aliyun-net-sdk-core-unit-tests", "aliyun-net-sdk-core.Tests\aliyun-net-sdk-core-unit-tests.csproj", "{5E4C70FD-F788-40BD-B313-832E30ED91BF}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "aliyun-net-feature-test", "aliyun-net-feature-test\aliyun-net-feature-test.csproj", "{546277AB-F1B7-4F9F-A4F2-45000AAE6F57}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -44,5 +46,17 @@ Global
 		{5E4C70FD-F788-40BD-B313-832E30ED91BF}.Release|x64.Build.0 = Release|Any CPU
 		{5E4C70FD-F788-40BD-B313-832E30ED91BF}.Release|x86.ActiveCfg = Release|Any CPU
 		{5E4C70FD-F788-40BD-B313-832E30ED91BF}.Release|x86.Build.0 = Release|Any CPU
+		{546277AB-F1B7-4F9F-A4F2-45000AAE6F57}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{546277AB-F1B7-4F9F-A4F2-45000AAE6F57}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{546277AB-F1B7-4F9F-A4F2-45000AAE6F57}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{546277AB-F1B7-4F9F-A4F2-45000AAE6F57}.Debug|x64.Build.0 = Debug|Any CPU
+		{546277AB-F1B7-4F9F-A4F2-45000AAE6F57}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{546277AB-F1B7-4F9F-A4F2-45000AAE6F57}.Debug|x86.Build.0 = Debug|Any CPU
+		{546277AB-F1B7-4F9F-A4F2-45000AAE6F57}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{546277AB-F1B7-4F9F-A4F2-45000AAE6F57}.Release|Any CPU.Build.0 = Release|Any CPU
+		{546277AB-F1B7-4F9F-A4F2-45000AAE6F57}.Release|x64.ActiveCfg = Release|Any CPU
+		{546277AB-F1B7-4F9F-A4F2-45000AAE6F57}.Release|x64.Build.0 = Release|Any CPU
+		{546277AB-F1B7-4F9F-A4F2-45000AAE6F57}.Release|x86.ActiveCfg = Release|Any CPU
+		{546277AB-F1B7-4F9F-A4F2-45000AAE6F57}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
